Less special casing for dacpac reverse engineering
