+++ src/GUI/RevEng.Core.Abstractions/RevEng.Core.Abstractions.csproj
@@ -4,6 +4,9 @@
     <TargetFramework>netstandard2.0</TargetFramework>
     <NoWarn>$(NoWarn);EF1001</NoWarn>
     <RootNamespace>RevEng.Core.Interfaces</RootNamespace>
+    <SignAssembly>true</SignAssembly>
+    <AssemblyOriginatorKeyFile>key.snk</AssemblyOriginatorKeyFile>
+    <DelaySign>false</DelaySign>
   </PropertyGroup>
 
 </Project>
