+++ src/GUI/RevEng.Core/Procedures/SqlServerStoredProcedureModelFactory.cs
@@ -8,7 +8,6 @@ using RevEng.Core.Abstractions.Model;
 using System;
 using System.Collections.Generic;
 using System.Data;
-using System.IO;
 using System.Linq;
 
 namespace RevEng.Core.Procedures
@@ -22,24 +21,13 @@ namespace RevEng.Core.Procedures
             _logger = logger;
         }
 
-        public ProcedureModel Create(string dacpacPath, ProcedureModelFactoryOptions options)
+        public ProcedureModel Create(string connectionString, ProcedureModelFactoryOptions options)
         {
-            if (string.IsNullOrEmpty(dacpacPath))
-            {
-                throw new ArgumentException(@"invalid path", nameof(dacpacPath));
-            }
-            if (!File.Exists(dacpacPath))
-            {
-                throw new ArgumentException($"Dacpac file not found: {dacpacPath}");
-            }
-            return GetStoredProcedures(dacpacPath, options);
+            return GetStoredProcedures(connectionString, options);
         }
 
         private ProcedureModel GetStoredProcedures(string connectionString, ProcedureModelFactoryOptions options)
         {
-
-
-
             var dtResult = new DataTable();
             var result = new List<Procedure>();
             var errors = new List<string>();
